fix: separate chess and piece class
chess 클래스와 piece 클래스를 분할
